==> src/12_mvt_prepare_stat_data.ipynb <==
**Source Code Authors :**   
> *Abigail Swamidoss (Abigail.Swamidoss@gmail.com), Samhith Kethireddy (Kethireddy.samhith@gmail.com)*  

**Published in Research Article:**  
> [Computational Analysis of Routine Biopsies Improves Diagnosis and Prediction of Cardiac Allograft Vasculopathy](https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.121.058459)  
> *by Eliot G. Peyster, Andrew Janowczyk, Abigail Swamidoss, Samhith Kethireddy, Michael D. Feldman and Kenneth B. Margulies*  
*Originally published 11 Apr 2022*  
  
> [Supplementary Material](https://www.ahajournals.org/action/downloadSupplement?doi=10.1161%2FCIRCULATIONAHA.121.058459&file=10.1161.circulationaha.121.058459_supplemental_materials.pdf)  

**Publisher**
[Journal of American Heart Association (JAHA) - Circulation](https://www.ahajournals.org/journal/circ)

#  Install, Setup gdrive and imports

import os
import pandas as pd


IN_CWR_SERVER = False 
slideset = 'cav2'

# Files


if IN_CWR_SERVER : # CWR server
  datadir = "/directory_in_CWR_server/"
else : # LOCAL
  datadir = "C:\\research\\cav\\datamvt\\"

# Input Folders and Files
outdir = datadir + "output\\" + slideset + "\\"
dc_f = outdir + "11_mvt_DC.csv.zip"
hc_f = outdir + "11_mvt_HC.csv.zip" 
dy_f = outdir + "11_mvt_DY.csv.zip" 
us_f = outdir + "11_mvt_US.csv.zip" 
ma_f = outdir + "11_mvt_MA.csv.zip" 
dc_os_f = outdir + "11_mvt_DC_OS.csv.zip" 
hc_os_f = outdir + "11_mvt_HC_OS.csv.zip" 
dy_os_f = outdir + "11_mvt_DY_OS.csv.zip" 
us_os_f = outdir + "11_mvt_US_OS.csv.zip" 
ma_os_f = outdir + "11_mvt_MA_OS.csv.zip"

#output folder and files
statdir = datadir + "stats\\" + slideset + "\\"
outfname = '12_movat_all.csv'


# Main

raw_dc_df = pd.DataFrame()
raw_hc_df = pd.DataFrame()
raw_dy_df = pd.DataFrame()
raw_us_df = pd.DataFrame()
raw_ma_df = pd.DataFrame()
raw_dc_os_df = pd.DataFrame()
raw_hc_os_df = pd.DataFrame()
raw_dy_os_df = pd.DataFrame()
raw_us_os_df = pd.DataFrame()
raw_ma_os_df = pd.DataFrame()

if os.path.exists(dc_f):
    raw_dc_df = pd.read_csv(dc_f) 
if os.path.exists(hc_f):
    raw_hc_df = pd.read_csv(hc_f) 
if os.path.exists(dy_f):
    raw_dy_df = pd.read_csv(dy_f) 
if os.path.exists(us_f):
    raw_us_df = pd.read_csv(us_f) 
if os.path.exists(ma_f):
    raw_ma_df = pd.read_csv(ma_f) 

if os.path.exists(dc_os_f):
    raw_dc_os_df = pd.read_csv(dc_os_f) 
if os.path.exists(hc_os_f):
    raw_hc_os_df = pd.read_csv(hc_os_f) 
if os.path.exists(dy_os_f):
    raw_dy_os_df = pd.read_csv(dy_os_f) 
if os.path.exists(us_os_f):
    raw_us_os_df = pd.read_csv(us_os_f) 
if os.path.exists(ma_os_f):
    raw_ma_os_df = pd.read_csv(ma_os_f) 

comb_df = pd.concat([raw_dc_df, raw_hc_df, raw_dy_df, raw_us_df, raw_ma_df, raw_dc_os_df, raw_hc_os_df, raw_dy_os_df, raw_us_os_df, raw_ma_os_df], ignore_index=True)
#comb_df.to_csv(statdir +"movat_all_test0.csv", index=True)

GROUP_BY= ['cohort','fname']
COLUMNS=['black', 'white', 'blue', 'collagen', 'myocyte', 'myocar', 'stro_myocar', 'collagen_myocar', 'blue_myocar', 'white_myocar']
#comb_df = comb_df.groupby(['cohort','fname'])[['black', 'white', 'blue', 'collagen', 'myocyte', 'myocardium']].sum()
comb_df = comb_df.groupby(GROUP_BY)[COLUMNS].sum()

#comb_df.to_csv(outdir +"movat_all.csv", index=True)
comb_df.to_csv(f"{outdir}{outfname}.zip", index=True, compression=dict(method='zip', archive_name=f'{outfname}'))
print(f"output saved in {outdir}{outfname}.zip")
