==> src/32_cav_add_computed_cols.ipynb <==
**Source Code Authors :**   
> *Abigail Swamidoss (Abigail.Swamidoss@gmail.com), Samhith Kethireddy (Kethireddy.samhith@gmail.com)*  

**Published in Research Article:**  
> [Computational Analysis of Routine Biopsies Improves Diagnosis and Prediction of Cardiac Allograft Vasculopathy](https://www.ahajournals.org/doi/10.1161/CIRCULATIONAHA.121.058459)  
> *by Eliot G. Peyster, Andrew Janowczyk, Abigail Swamidoss, Samhith Kethireddy, Michael D. Feldman and Kenneth B. Margulies*  
*Originally published 11 Apr 2022*  
  
> [Supplementary Material](https://www.ahajournals.org/action/downloadSupplement?doi=10.1161%2FCIRCULATIONAHA.121.058459&file=10.1161.circulationaha.121.058459_supplemental_materials.pdf)  

**Publisher**
[Journal of American Heart Association (JAHA) - Circulation](https://www.ahajournals.org/journal/circ)

#  Install, Setup gdrive and imports

import os
import pandas as pd

IN_CWR_SERVER = False 
slideset = 'cav2'

# Files


from pathlib import Path
if IN_CWR_SERVER :
  DATADIR = "/gdrive/My Drive/PytorchDigitalPathology/qupath/data/"
else :
  projectdir = "C:\\research\\cav\\"
  DATADIR = projectdir + "datacd31\\"

#input folders and files
outdir = DATADIR + "output\\" + slideset + "\\"
CD31_MVT_DIL_CSV = "31cd31_mvt_dilated.csv"
CD31_MVT_UDL_CSV = "31cd31_mvt_undilated.csv"

#output folders and files
STATDIR = projectdir + "stats\\" + slideset + "\\"
FINAL_DIL_CSV = "32_final_dilated.csv"
FINAL_UDL_CSV = "32_final_undilated.csv"

HEALTHY = ['HC', 'US', 'MA']
DESEASED = ['DC']     
DESEASED_YR1 = ['DY'] 
COHORTS = HEALTHY + DESEASED + DESEASED_YR1


# Functions

def addComputedColumnsInPlace(df) :
    df.eval('myocyte_div_tissue = myocyte/tissue', inplace=True)
    df.eval('blue_myocar_collagen_myocar_div_myocyte = (blue_myocar + collagen_myocar)/myocyte', inplace=True)
    df.eval('in_dab_div_dab_area = in_dab/dab_area', inplace=True)
    #bin area / bin count
    df.eval('qbin1area_div_count = dab_q1_area/dab_q1', inplace=True)
    df.eval('qbin2area_div_count = dab_q2_area/dab_q2', inplace=True)
    df.eval('qbin3area_div_count = dab_q3_area/dab_q3', inplace=True)
    df.eval('qbin4area_div_count = dab_q4_area/dab_q4', inplace=True)

    df.eval('sbin1area_div_count = dab_s1_area/dab_s1', inplace=True)
    df.eval('sbin2area_div_count = dab_s2_area/dab_s2', inplace=True)
    df.eval('sbin3area_div_count = dab_s3_area/dab_s3', inplace=True)
    df.eval('sbin4area_div_count = dab_s4_area/dab_s4', inplace=True)
    df.eval('sbin5area_div_count = dab_s5_area/dab_s5', inplace=True)
    df.eval('sbin6area_div_count = dab_s6_area/dab_s6', inplace=True)

    df.eval('bbin1area_div_count = dab_b1_area/dab_b1', inplace=True)
    df.eval('bbin2area_div_count = dab_b2_area/dab_b2', inplace=True)
    df.eval('bbin3area_div_count = dab_b3_area/dab_b3', inplace=True)
    df.eval('bbin4area_div_count = dab_b4_area/dab_b4', inplace=True)
    df.eval('bbin5area_div_count = dab_b5_area/dab_b5', inplace=True)

    #bin area / total dab area
    df.eval('qbin1area_div_dab_area = dab_q1_area/dab_area', inplace=True)
    df.eval('qbin2area_div_dab_area = dab_q2_area/dab_area', inplace=True)
    df.eval('qbin3area_div_dab_area = dab_q3_area/dab_area', inplace=True)
    df.eval('qbin4area_div_dab_area = dab_q4_area/dab_area', inplace=True)

    df.eval('sbin1area_div_dab_area = dab_s1_area/dab_area', inplace=True)
    df.eval('sbin2area_div_dab_area  = dab_s2_area/dab_area', inplace=True)
    df.eval('sbin3area_div_dab_area  = dab_s3_area/dab_area', inplace=True)
    df.eval('sbin4area_div_dab_area  = dab_s4_area/dab_area', inplace=True)
    df.eval('sbin5area_div_dab_area  = dab_s5_area/dab_area', inplace=True)
    df.eval('sbin6area_div_dab_area  = dab_s6_area/dab_area', inplace=True)

    df.eval('bbin1area_div_dab_area  = dab_b1_area/dab_area', inplace=True)
    df.eval('bbin2area_div_dab_area = dab_b2_area/dab_area', inplace=True)
    df.eval('bbin3area_div_dab_area  = dab_b3_area/dab_area', inplace=True)
    df.eval('bbin4area_div_dab_area  = dab_b4_area/dab_area', inplace=True)
    df.eval('bbin5area_div_dab_area  = dab_b5_area/dab_area', inplace=True)

    #bin area / myocar
    df.eval('qbin1area_div_myocar = dab_q1_area/myocar', inplace=True)
    df.eval('qbin2area_div_myocar = dab_q2_area/myocar', inplace=True)
    df.eval('qbin3area_div_myocar = dab_q3_area/myocar', inplace=True)
    df.eval('qbin4area_div_myocar = dab_q4_area/myocar', inplace=True)

    df.eval('sbin1area_div_myocar = dab_s1_area/myocar', inplace=True)
    df.eval('sbin2area_div_myocar = dab_s2_area/myocar', inplace=True)
    df.eval('sbin3area_div_myocar = dab_s3_area/myocar', inplace=True)
    df.eval('sbin4area_div_myocar = dab_s4_area/myocar', inplace=True)
    df.eval('sbin5area_div_myocar = dab_s5_area/myocar', inplace=True)
    df.eval('sbin6area_div_myocar = dab_s6_area/myocar', inplace=True)

    df.eval('bbin1area_div_myocar = dab_b1_area/myocar', inplace=True)
    df.eval('bbin2area_div_myocar = dab_b2_area/myocar', inplace=True)
    df.eval('bbin3area_div_myocar = dab_b3_area/myocar', inplace=True)
    df.eval('bbin4area_div_myocar = dab_b4_area/myocar', inplace=True)
    df.eval('bbin5area_div_myocar = dab_b5_area/myocar', inplace=True)

    #bin area / myocyte
    df.eval('qbin1area_div_myocyte = dab_q1_area/myocyte', inplace=True)
    df.eval('qbin2area_div_myocyte = dab_q2_area/myocyte', inplace=True)
    df.eval('qbin3area_div_myocyte = dab_q3_area/myocyte', inplace=True)
    df.eval('qbin4area_div_myocyte = dab_q4_area/myocyte', inplace=True)

    df.eval('sbin1area_div_myocyte = dab_s1_area/myocyte', inplace=True)
    df.eval('sbin2area_div_myocyte = dab_s2_area/myocyte', inplace=True)
    df.eval('sbin3area_div_myocyte = dab_s3_area/myocyte', inplace=True)
    df.eval('sbin4area_div_myocyte = dab_s4_area/myocyte', inplace=True)
    df.eval('sbin5area_div_myocyte = dab_s5_area/myocyte', inplace=True)
    df.eval('sbin6area_div_myocyte = dab_s6_area/myocyte', inplace=True)

    df.eval('bbin1area_div_myocyte = dab_b1_area/myocyte', inplace=True)
    df.eval('bbin2area_div_myocyte = dab_b2_area/myocyte', inplace=True)
    df.eval('bbin3area_div_myocyte = dab_b3_area/myocyte', inplace=True)
    df.eval('bbin4area_div_myocyte = dab_b4_area/myocyte', inplace=True)
    df.eval('bbin5area_div_myocyte = dab_b5_area/myocyte', inplace=True)

    return

def addAddlComputedColumnsInPlace(df) :
    df.eval('nuc_in_dab_div_tot_nuc_count = in_dab/nuc_count', inplace=True)
    df.eval('nuc_in_dab_div_nuc_outside = in_dab/(nuc_count-in_dab)', inplace=True)
    df.eval('nuc_in_dab_div_dab_area = in_dab/dab_area', inplace=True)
    df.eval('nuc_in_dab_div_hem_area = in_dab/myocyte', inplace=True)
    df.eval('nuc_in_dab_div_tis_area = in_dab/tissue', inplace=True)
    df.eval('dab_area_div_hem_area = dab_area/myocyte', inplace=True)
    df.eval('dab_area_div_tis_area = dab_area/tissue', inplace=True)
    df.eval('dab_area_div_tot_nuc_count = dab_area/nuc_count', inplace=True)
    df.eval('dab_area_div_nuc_outside = dab_area/(nuc_count-in_dab)', inplace=True)
    df.eval('tot_nuc_count_div_tis_area = nuc_count/tissue', inplace=True)
    df.eval('tot_nuc_count_div_hem_area = nuc_count/myocyte', inplace=True)
    df.eval('nuc_outside_div_tis_area = (nuc_count-in_dab)/tissue', inplace=True)
    df.eval('nuc_outside_div_hem_area = (nuc_count-in_dab)/myocyte', inplace=True)
    df.eval('tot_nuc_count_div_myocyte = nuc_count/myocyte', inplace=True)
    df.eval('tis_area_div_myocyte = tissue/myocyte', inplace=True)
    df.eval('dab_area_div_myocyte = dab_area/myocyte', inplace=True)
    df.eval('collagen_div_myocyte = collagen/myocyte', inplace=True)
    df.eval('blue_div_myocyte = blue/myocyte', inplace=True)
    df.eval('myocyte_div_myocar = myocyte/myocar', inplace=True)
    df.eval('collagen_myocar_div_myocar = collagen_myocar/myocar', inplace=True)
    df.eval('stro_myocar_div_myocar = stro_myocar/myocar', inplace=True)
    df.eval('blue_myocar_div_myocar = blue_myocar/myocar', inplace=True)
    return

# Main 

## Merge CD31 with Movat

if os.path.exists(f"{outdir}{CD31_MVT_DIL_CSV}.zip"):
    print(f"loading {outdir}{CD31_MVT_DIL_CSV}.zip")
    dil_df=pd.read_csv(f"{outdir}{CD31_MVT_DIL_CSV}.zip")
    addComputedColumnsInPlace(dil_df)
    addAddlComputedColumnsInPlace(dil_df)
    print(f'saving {STATDIR}{FINAL_DIL_CSV}.zip')
    dil_df.to_csv(f"{STATDIR}{FINAL_DIL_CSV}.zip", index=False, compression=dict(method='zip', archive_name=f'{FINAL_DIL_CSV}'))
else :
    print(f"unable to load not load {outdir}{CD31_MVT_DIL_CSV}.zip" )

if os.path.exists(f"{outdir}{CD31_MVT_UDL_CSV}.zip"):
    print(f"loading {outdir}{CD31_MVT_UDL_CSV}.zip")
    udil_df=pd.read_csv(f"{outdir}{CD31_MVT_UDL_CSV}.zip")
    addComputedColumnsInPlace(udil_df)
    addAddlComputedColumnsInPlace(udil_df)
    print(f'saving {STATDIR}{FINAL_UDL_CSV}.zip')
    udil_df.to_csv(f"{STATDIR}{FINAL_UDL_CSV}.zip", index=False, compression=dict(method='zip', archive_name=f'{FINAL_UDL_CSV}'))
